Separate the cover styles from the main app. css
I'm not sure why, but the body styles specified
in cover.css were applying on the project index
page (e.g. the box-shadow), even though
cover.css was not supposed to be included on that
page. This is weird, because the line in the 
cover layout that includes cover.css explicitly
was definitely having an effect. Maybe all styles
were being loaded, and what I was actually doing
was changing the order in which the rules were
being applied.