Streamline the github oauth process
Instead of waiting for the user to choose to 
connect their github account, simply start the
connection process as soon as they do something 
which requires it.

We stash the target repo name in the session, so
we can redirect the user to the correct page when
they have completed the authorisation step.